tests: parallelise 'make valgrind'
Use $(MAKE) instead of 'make' inside the Makefile,
otherwise the make will run as if -j1 had been
specified and complain about the job server not
being available, and with $(MAKE) in inherits the
parent make's settings it seems.

Upgrade common submodule for parallel check-valgrind.

Let this settle a bit before upgrading the other modules.