Revert "openh264: fix to work with released versions and master of upstream"
This reverts commit 15394aa70543f55c54fa8a474805224f4138f76c.

The latest release (v1.1) does not have pkg-config support
yet, so this plugin won't be built with the latest release.
Cerbero uses the latest release, so this makes cerbero
builds fail, which expect the plugin to be built.

We can re-commit this once there's a release that includes
pkg-config support.